can hide short flags via config file

+++ features/steps/common.rb
@@ -8,11 +8,14 @@ def in_home_folder(&block)
 end
 
 Given %r{^a safe folder} do
-  FileUtils.rm_rf   @tmp_root = File.dirname(__FILE__) + "/../../tmp"
-  FileUtils.mkdir_p @tmp_root
-  FileUtils.mkdir_p @home_path = File.expand_path(File.join(@tmp_root, "home"))
-  @lib_path = File.expand_path(File.dirname(__FILE__) + '/../../lib')
-  Given "env variable $HOME set to '#{@home_path}'"
+  unless @have_created_safe_folder
+    FileUtils.rm_rf   @tmp_root = File.dirname(__FILE__) + "/../../tmp"
+    FileUtils.mkdir_p @tmp_root
+    FileUtils.mkdir_p @home_path = File.expand_path(File.join(@tmp_root, "home"))
+    @lib_path = File.expand_path(File.dirname(__FILE__) + '/../../lib')
+    Given "env variable $HOME set to '#{@home_path}'"
+    @have_created_safe_folder = true
+  end
 end
 
 Given %r{^this project is active project folder} do
